+++ lib/selecto/db/adapter.ex
@@ -55,6 +55,9 @@ defmodule Selecto.DB.Adapter do
   @callback list_tables(connection(), introspection_options()) ::
               {:ok, [String.t()]} | {:error, term()}
 
+  @callback list_relations(connection(), introspection_options()) ::
+              {:ok, [map()]} | {:error, term()}
+
   @callback introspect_table(connection(), String.t(), introspection_options()) ::
               {:ok, schema_metadata()} | {:error, term()}
 
@@ -77,6 +80,7 @@ defmodule Selecto.DB.Adapter do
                       execute_repo_fallback: 3,
                       start_pool: 3,
                       list_tables: 2,
+                      list_relations: 2,
                       introspect_table: 3,
                       format_datetime: 2,
                       rollup_sql: 1,